requirements.txt
python-dotenv
google-cloud-storage
google-cloud-logging

#En MACOS
psycopg2-binary

#CLI
typer

#
fastapi


